Refactor: Split tests into unit and integration
Moves the existing integration test suite to a new `integration_test.go` file.
Retains the unit test suites (Config, parseAddress, parseWGKey) in `main_test.go`.
Cleans up unused imports in both files to reflect their specific test scopes.

This separation improves test organization and clarity.
